requirements.txt
python-dotenv
pypandoc
google-genai
